refactor: remove redundant line_sentence_chunker directory
