+++ src/ingestion_pipeline/chunking/document_chunker_factory.py
@@ -6,9 +6,9 @@ from ingestion_pipeline.chunking.base_document_chunker import DocumentChunker
 from ingestion_pipeline.chunking.layout_handler.layout_chunking_config import LayoutChunkingConfig
 from ingestion_pipeline.chunking.layout_handler.strategies.key_value.layout_key_value import KeyValueChunker
 from ingestion_pipeline.chunking.layout_handler.strategies.layout_text import LayoutTextChunkingStrategy
-from ingestion_pipeline.chunking.layout_handler.strategies.line_sentence_chunker import LineSentenceChunkingConfig
 from ingestion_pipeline.chunking.layout_handler.strategies.list.list_chunker import LayoutListChunkingStrategy
 from ingestion_pipeline.chunking.layout_handler.strategies.table.layout_table import LayoutTableChunkingStrategy
+from ingestion_pipeline.chunking.line_sentence_chunker.config import LineSentenceChunkingConfig
 from ingestion_pipeline.chunking.line_sentence_splitter import LineBasedDocumentChunker
 from ingestion_pipeline.chunking.textract_layout_chunker import TextractLayoutDocumentChunker
 from ingestion_pipeline.config import settings
