+++ src/ingestion_pipeline/chunking/line_sentence_splitter.py
@@ -12,10 +12,8 @@ from textractor.entities.document import Document
 
 from ingestion_pipeline.chunking.base_document_chunker import ChunkError, DocumentChunker
 from ingestion_pipeline.chunking.exceptions import ChunkException
-from ingestion_pipeline.chunking.layout_handler.strategies.line_sentence_chunker import (
-    LineSentenceChunker,
-    LineSentenceChunkingConfig,
-)
+from ingestion_pipeline.chunking.line_sentence_chunker.chunker import LineSentenceChunker
+from ingestion_pipeline.chunking.line_sentence_chunker.config import LineSentenceChunkingConfig
 from ingestion_pipeline.chunking.schemas import DocumentChunk, DocumentMetadata, ProcessedDocument
 from ingestion_pipeline.config import settings
